requirements.txt
pytest
requests
PyHamcrest
allure-pytest
